Move date functions into dateUtils
